refactor and update requirements
