+++ CMakeLists.txt
@@ -96,16 +96,6 @@ else()
     message(STATUS "half.hpp is at ${HALF_INCLUDE_DIR}")
 endif()
 
-if(MIGRAPHX_USE_BLAZE)
-    find_path(BLAZE_INCLUDE_DIR blaze/Blaze.h)
-    if(NOT BLAZE_INCLUDE_DIR)
-        message(STATUS "Blaze not found, disabling MIGRAPHX_USE_BLAZE")
-        set(MIGRAPHX_USE_BLAZE OFF)
-    else()
-        message(STATUS "Blaze found at ${BLAZE_INCLUDE_DIR}")
-    endif()
-endif()
-
 include(CheckTypeSize)
 set(CMAKE_REQUIRED_INCLUDES ${HALF_INCLUDE_DIR})
 set(CMAKE_EXTRA_INCLUDE_FILES half.hpp)
