+++ src/include/migraphx/gemm.hpp
@@ -25,270 +25,12 @@
 #define MIGRAPHX_GUARD_RTGLIB_GEMM_HPP
 
 #include <migraphx/config.hpp>
-#include <migraphx/dfor.hpp>
-#include <migraphx/par_for.hpp>
-#include <migraphx/tensor_view.hpp>
-#include <migraphx/float_equal.hpp>
-#include <migraphx/env.hpp>
-
-#if MIGRAPHX_USE_BLAZE
-#include <blaze/Blaze.h>
-#include <vector>
-#endif
+#include <migraphx/argument.hpp>
 
 namespace migraphx {
 inline namespace MIGRAPHX_INLINE_NS {
 
-MIGRAPHX_DECLARE_ENV_VAR(MIGRAPHX_BLAZE_DEBUG)
-MIGRAPHX_DECLARE_ENV_VAR(MIGRAPHX_DISABLE_BLAZE)
-
-#if MIGRAPHX_USE_BLAZE
-template <class T>
-using blaze_row_major = blaze::CustomMatrix<T, blaze::unaligned, blaze::unpadded, blaze::rowMajor>;
-template <class T>
-using blaze_col_major =
-    blaze::CustomMatrix<T, blaze::unaligned, blaze::unpadded, blaze::columnMajor>;
-
-template <class T>
-constexpr bool is_blaze_native_type()
-{
-    return std::is_same<T, float>{} or std::is_same<T, double>{};
-}
-
-enum class mat_order
-{
-    row_major,
-    col_major,
-    unsupported
-};
-
-struct blaze_layout
-{
-    mat_order order     = mat_order::unsupported;
-    std::size_t spacing = 0;
-    bool packed         = false;
-};
-
-inline blaze_layout get_blaze_layout(const shape& s)
-{
-    if(s.ndim() < 2)
-        return {};
-    auto dim_0 = s.ndim() - 2;
-    auto dim_1 = s.ndim() - 1;
-    auto s0    = s.strides()[dim_0];
-    auto s1    = s.strides()[dim_1];
-    auto rows  = s.lens()[dim_0];
-    auto cols  = s.lens()[dim_1];
-    if(s1 == 1 and s0 >= cols)
-        return {mat_order::row_major, s0, s0 == cols};
-    if(s0 == 1 and s1 >= rows)
-        return {mat_order::col_major, s1, s1 == rows};
-    return {};
-}
-
-inline std::size_t batch_offset(const shape& s, std::size_t batch, std::size_t n_batch_dims)
-{
-    std::size_t offset    = 0;
-    std::size_t remaining = batch;
-    for(std::size_t d = n_batch_dims; d > 0; d--)
-    {
-        offset += (remaining % s.lens()[d - 1]) * s.strides()[d - 1];
-        remaining /= s.lens()[d - 1];
-    }
-    return offset;
-}
-
-template <class T, class Func>
-void with_blaze_mat(T* ptr, std::size_t rows, std::size_t cols, mat_order order, Func&& func)
-{
-    if(order == mat_order::col_major)
-    {
-        blaze_col_major<T> mat(ptr, rows, cols);
-        func(mat);
-    }
-    else
-    {
-        blaze_row_major<T> mat(ptr, rows, cols);
-        func(mat);
-    }
-}
-
-template <class T, class U>
-void copy_2d(T* dst,
-             std::size_t dst_row_stride,
-             std::size_t dst_col_stride,
-             const U* src,
-             std::size_t src_row_stride,
-             std::size_t src_col_stride,
-             std::size_t rows,
-             std::size_t cols)
-{
-    for(std::size_t i = 0; i < rows; i++)
-        for(std::size_t j = 0; j < cols; j++)
-            dst[i * dst_row_stride + j * dst_col_stride] =
-                static_cast<T>(src[i * src_row_stride + j * src_col_stride]);
-}
-#endif
-
-template <class T, class U, class F>
-void gemm(tensor_view<T> cmat, tensor_view<U> amat, tensor_view<U> bmat, F alpha, F beta)
-{
-    std::size_t n_dims = cmat.get_shape().lens().size();
-    std::size_t dim_0  = n_dims - 2;
-    std::size_t dim_1  = n_dims - 1;
-    auto k             = amat.get_shape().lens()[dim_1];
-
-    assert(amat.get_shape().lens()[dim_1] == bmat.get_shape().lens()[dim_0]);
-    assert(cmat.get_shape().lens()[dim_0] == amat.get_shape().lens()[dim_0]);
-    assert(cmat.get_shape().lens()[dim_1] == bmat.get_shape().lens()[dim_1]);
-
-#if MIGRAPHX_USE_BLAZE
-    if(not enabled(MIGRAPHX_DISABLE_BLAZE{}))
-    {
-        auto m_size = amat.get_shape().lens()[dim_0];
-        auto n_size = bmat.get_shape().lens()[dim_1];
-        auto k_size = k;
-
-        std::size_t num_batches = 1;
-        for(std::size_t i = 0; i < dim_0; i++)
-            num_batches *= cmat.get_shape().lens()[i];
-
-        // Native zero-copy path: float/double with packed row/col-major inner 2D
-        if constexpr(is_blaze_native_type<U>() and std::is_same<T, U>{})
-        {
-            auto a_layout = get_blaze_layout(amat.get_shape());
-            auto b_layout = get_blaze_layout(bmat.get_shape());
-            auto c_layout = get_blaze_layout(cmat.get_shape());
-
-            if(a_layout.packed and b_layout.packed and c_layout.packed)
-            {
-                if(enabled(MIGRAPHX_BLAZE_DEBUG{}))
-                {
-                    std::cerr << "[blaze gemm] " << m_size << "x" << k_size << "x" << n_size
-                              << " batches=" << num_batches << " (native)" << std::endl;
-                }
-
-                for(std::size_t batch = 0; batch < num_batches; batch++)
-                {
-                    auto a_off = batch_offset(amat.get_shape(), batch, dim_0);
-                    auto b_off = batch_offset(bmat.get_shape(), batch, dim_0);
-                    auto c_off = batch_offset(cmat.get_shape(), batch, dim_0);
-
-                    with_blaze_mat(
-                        amat.data() + a_off, m_size, k_size, a_layout.order, [&](auto& a_mat) {
-                            with_blaze_mat(bmat.data() + b_off,
-                                           k_size,
-                                           n_size,
-                                           b_layout.order,
-                                           [&](auto& b_mat) {
-                                               with_blaze_mat(cmat.data() + c_off,
-                                                              m_size,
-                                                              n_size,
-                                                              c_layout.order,
-                                                              [&](auto& c_mat) {
-                                                                  if(float_equal(alpha, F{1}) and
-                                                                     float_equal(beta, F{0}))
-                                                                      c_mat = a_mat * b_mat;
-                                                                  else
-                                                                      c_mat =
-                                                                          alpha * (a_mat * b_mat) +
-                                                                          beta * c_mat;
-                                                              });
-                                           });
-                        });
-                }
-                return;
-            }
-        }
-
-        // Copy path: works for any type and any layout.
-        // Copies each batch slice to standard row-major float buffers using raw strides,
-        // runs Blaze GEMM, then copies back.
-        auto a_row_stride = amat.get_shape().strides()[dim_0];
-        auto a_col_stride = amat.get_shape().strides()[dim_1];
-        auto b_row_stride = bmat.get_shape().strides()[dim_0];
-        auto b_col_stride = bmat.get_shape().strides()[dim_1];
-        auto c_row_stride = cmat.get_shape().strides()[dim_0];
-        auto c_col_stride = cmat.get_shape().strides()[dim_1];
-
-        if(enabled(MIGRAPHX_BLAZE_DEBUG{}))
-        {
-            std::cerr << "[blaze gemm] " << m_size << "x" << k_size << "x" << n_size
-                      << " batches=" << num_batches << " (copy to float)" << std::endl;
-        }
-
-        std::vector<float> a_buf(m_size * k_size);
-        std::vector<float> b_buf(k_size * n_size);
-        std::vector<float> c_buf(m_size * n_size);
-
-        for(std::size_t batch = 0; batch < num_batches; batch++)
-        {
-            auto a_off = batch_offset(amat.get_shape(), batch, dim_0);
-            auto b_off = batch_offset(bmat.get_shape(), batch, dim_0);
-            auto c_off = batch_offset(cmat.get_shape(), batch, dim_0);
-
-            copy_2d(a_buf.data(),
-                    k_size,
-                    std::size_t{1},
-                    amat.data() + a_off,
-                    a_row_stride,
-                    a_col_stride,
-                    m_size,
-                    k_size);
-            copy_2d(b_buf.data(),
-                    n_size,
-                    std::size_t{1},
-                    bmat.data() + b_off,
-                    b_row_stride,
-                    b_col_stride,
-                    k_size,
-                    n_size);
-            if(not float_equal(beta, F{0}))
-                copy_2d(c_buf.data(),
-                        n_size,
-                        std::size_t{1},
-                        cmat.data() + c_off,
-                        c_row_stride,
-                        c_col_stride,
-                        m_size,
-                        n_size);
-
-            blaze_row_major<float> a(a_buf.data(), m_size, k_size);
-            blaze_row_major<float> b(b_buf.data(), k_size, n_size);
-            blaze_row_major<float> c(c_buf.data(), m_size, n_size);
-
-            if(float_equal(alpha, F{1}) and float_equal(beta, F{0}))
-                c = a * b;
-            else
-                c = alpha * (a * b) + beta * c;
-
-            copy_2d(cmat.data() + c_off,
-                    c_row_stride,
-                    c_col_stride,
-                    c_buf.data(),
-                    n_size,
-                    std::size_t{1},
-                    m_size,
-                    n_size);
-        }
-        return;
-    }
-#endif
-
-    auto cs = cmat.get_shape();
-    par_for(cs.elements(), [&](auto i) {
-        auto c_idx = cs.multi(i);
-        auto a_idx = c_idx;
-        auto b_idx = c_idx;
-        double s   = 0.0;
-        dfor(k)([&](auto kk) {
-            a_idx[dim_1] = b_idx[dim_0] = kk;
-            s += static_cast<double>(amat(a_idx.begin(), a_idx.end())) *
-                 static_cast<double>(bmat(b_idx.begin(), b_idx.end()));
-        });
-        cmat(c_idx.begin(), c_idx.end()) = alpha * s + cmat(c_idx.begin(), c_idx.end()) * beta;
-    });
-}
+void gemm(const argument& c_arg, const argument& a_arg, const argument& b_arg);
 
 } // namespace MIGRAPHX_INLINE_NS
 } // namespace migraphx
